backend_fastapi/pyproject.toml
[project]
name = "collaborative-editor-ai-backend-fastapi"
version = "0.1.0"
description = "Assignment 2 FastAPI backend for the collaborative editor"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.4.0",
  "pytest>=8.3.0",
  "httpx>=0.27.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
